Update multi inline adaptive recycling example with comments and best practices
